refactor: reorganize types and improve route structure
- Refactor ProtectedRoute to use Outlet for nested routes pattern
- Split types into feature-specific files (auth.ts, transaction.ts)
- Keep only shared types (ApiResponse, User) in main types/index.ts
- Add missing route constants (LOGOUT, ME) for API endpoints
- Update all import paths to use new type file locations
