+++ frontend/src/pages/TransactionHistoryPage.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { ROUTES } from '@/constants'
 import { TRADE_TYPE, CURRENCY } from '@/constants'
-import type { Transaction } from '@/types'
+import type { Transaction } from '@/types/transaction'
 import { Card, CardContent } from '@/components/ui/card'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
