+++ frontend/src/types/auth.ts
@@ -1,0 +1,23 @@
+import type { User } from './index'
+
+export interface LoginRequest {
+  email: string
+  password: string
+}
+
+export interface LoginResponse {
+  success: boolean
+  message: string
+  data: {
+    token: string
+    user: User
+  }
+}
+
+export interface AuthState {
+  user: User | null
+  token: string | null
+  isAuthenticated: boolean
+  isLoading: boolean
+  error: string | null
+}
